Update dependencies, add intro, document more things.
This updates doca to use the latest json-schema-example-loader 2.0.0.
It also adds a "getting started" introduction section and a way
to hook it to the table of contents, and documents a few more
things in the config file.  The config file now starts uncommented,
as it has a reasonable default for the headers.